+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   * visualisation of map in ios 
   * support select position (custom marker, not network image will not work) 
   * support track user location
+## 0.7.10 :
+* (#123) fix zoom when changeLocation was called
+* (#123) remove previous marker  when changeLocation was called
 ## 0.7.9+3 :
 * remove background location permission
 ## 0.7.9+2 :